reverting from Decimal to float

+++ py/munits/converters.py
@@ -1,5 +1,5 @@
 # encoding: utf-8
-from decimal import Decimal
+# from decimal import float
 from munits.unitnotation import UnitNotation
 
 
@@ -15,22 +15,22 @@ class Converter:
             return conv / (self._v ** exponent)
 
     prefixes = {
-        "E": ConverterFunction(Decimal("1e18")),  # exa
-        "P": ConverterFunction(Decimal("1e15")),  # peta
-        "T": ConverterFunction(Decimal("1e12")),  # tera
-        "G": ConverterFunction(Decimal("1e9")),  # giga
-        "M": ConverterFunction(Decimal("1e6")),  # mega
-        "k": ConverterFunction(Decimal("1e3")),  # kilo
-        "h": ConverterFunction(Decimal("1e2")),  # hecto
-        "da": ConverterFunction(Decimal("1e1")),  # deca
-        "d": ConverterFunction(Decimal("0.1")),  # deci
-        "c": ConverterFunction(Decimal("0.01")),  # centi
-        "m": ConverterFunction(Decimal("0.001")),  # milli
-        "μ": ConverterFunction(Decimal("0.000001")),  # micro
-        "n": ConverterFunction(Decimal("0.000000001")),  # nano
-        "p": ConverterFunction(Decimal("0.000000000001")),  # pico
-        "f": ConverterFunction(Decimal("0.000000000000001")),  # femto
-        "a": ConverterFunction(Decimal("0.000000000000000001"))  # atto
+        "E": ConverterFunction(float("1e18")),  # exa
+        "P": ConverterFunction(float("1e15")),  # peta
+        "T": ConverterFunction(float("1e12")),  # tera
+        "G": ConverterFunction(float("1e9")),  # giga
+        "M": ConverterFunction(float("1e6")),  # mega
+        "k": ConverterFunction(float("1e3")),  # kilo
+        "h": ConverterFunction(float("1e2")),  # hecto
+        "da": ConverterFunction(float("1e1")),  # deca
+        "d": ConverterFunction(float("0.1")),  # deci
+        "c": ConverterFunction(float("0.01")),  # centi
+        "m": ConverterFunction(float("0.001")),  # milli
+        "μ": ConverterFunction(float("0.000001")),  # micro
+        "n": ConverterFunction(float("0.000000001")),  # nano
+        "p": ConverterFunction(float("0.000000000001")),  # pico
+        "f": ConverterFunction(float("0.000000000000001")),  # femto
+        "a": ConverterFunction(float("0.000000000000000001"))  # atto
     }
 
     remove_prefix = []
@@ -44,7 +44,7 @@ class Converter:
                 if key not in self.remove_prefix:
                     self.units[key + self.base_unit] = value
 
-            self.units[self.base_unit] = Converter.ConverterFunction(Decimal("1.0"))
+            self.units[self.base_unit] = Converter.ConverterFunction(float("1.0"))
             if additional_units:
                 for key, value in additional_units.items():
                     self.units[key] = value
@@ -68,10 +68,10 @@ class LengthConvert(Converter):
         if not additional_units:
             additional_units = {}
         Converter.__init__(self, base_unit, {**{
-            "in": Converter.ConverterFunction(Decimal('0.0254')),
-            "ft": Converter.ConverterFunction(Decimal('0.3048')),
-            "yd": Converter.ConverterFunction(Decimal('0.914')),
-            "mi": Converter.ConverterFunction(Decimal('1609.344'))
+            "in": Converter.ConverterFunction(float('0.0254')),
+            "ft": Converter.ConverterFunction(float('0.3048')),
+            "yd": Converter.ConverterFunction(float('0.914')),
+            "mi": Converter.ConverterFunction(float('1609.344'))
         },
                                              **additional_units})
 
@@ -82,11 +82,11 @@ class MassConvert(Converter):
             additional_units = {}
 
         Converter.__init__(self, base_unit, {**{
-            "oz": Converter.ConverterFunction(Decimal('28.3495')),
-            "lb": Converter.ConverterFunction(Decimal('453.592')),
-            "t": Converter.ConverterFunction(Decimal('1e6')),
-            "tonne_uk": Converter.ConverterFunction(Decimal('1016046 ')),
-            "tonne_us": Converter.ConverterFunction(Decimal('907184. ')),
+            "oz": Converter.ConverterFunction(float('28.3495')),
+            "lb": Converter.ConverterFunction(float('453.592')),
+            "t": Converter.ConverterFunction(float('1e6')),
+            "tonne_uk": Converter.ConverterFunction(float('1016046 ')),
+            "tonne_us": Converter.ConverterFunction(float('907184. ')),
         },
                                              **additional_units})
 
@@ -98,9 +98,9 @@ class TimeConvert(Converter):
         if not additional_units:
             additional_units = {}
         Converter.__init__(self, base_unit, {**{
-            "min": Converter.ConverterFunction(Decimal('60')),
-            "h": Converter.ConverterFunction(Decimal('3600')),
-            "d": Converter.ConverterFunction(Decimal('86400'))
+            "min": Converter.ConverterFunction(float('60')),
+            "h": Converter.ConverterFunction(float('3600')),
+            "d": Converter.ConverterFunction(float('86400'))
         }, **additional_units})
 
 
@@ -122,15 +122,15 @@ class TemperatureConvert(Converter):
         if not additional_units:
             additional_units = {}
         Converter.__init__(self, base_unit, {**{
-            "K": TemperatureConvert.TempConverterFunction(Decimal('1.0'), Decimal('0.0')),
-            "°C": TemperatureConvert.TempConverterFunction(Decimal('1.0'), Decimal('273.15')),
-            "°F": TemperatureConvert.TempConverterFunction(Decimal('5.') / Decimal('9.'), Decimal('459.67'))
+            "K": TemperatureConvert.TempConverterFunction(float('1.0'), float('0.0')),
+            "°C": TemperatureConvert.TempConverterFunction(float('1.0'), float('273.15')),
+            "°F": TemperatureConvert.TempConverterFunction(float('5.') / float('9.'), float('459.67'))
         }, **additional_units})
 
 
 class AreaConvert(LengthConvert):
-    additional_units = {"ac": Converter.ConverterFunction(Decimal('4046.85')),
-                        "are": Converter.ConverterFunction(Decimal('100'))
+    additional_units = {"ac": Converter.ConverterFunction(float('4046.85')),
+                        "are": Converter.ConverterFunction(float('100'))
 
                         }
 
@@ -138,9 +138,9 @@ class AreaConvert(LengthConvert):
 class VolumeConvert(LengthConvert):
     def __init__(self, base_unit):
         LengthConvert.__init__(self, base_unit, {
-            "gal": Converter.ConverterFunction(Decimal('0.1558491279125191126493914617')),
-            "oz": Converter.ConverterFunction(Decimal('2.957e-5')),
-            "l": Converter.ConverterFunction(Decimal('1e-1'))
+            "gal": Converter.ConverterFunction(float('0.1558491279125191126493914617')),
+            "oz": Converter.ConverterFunction(float('2.957e-5')),
+            "l": Converter.ConverterFunction(float('1e-1'))
         }
                                )
